added reusable dollar input component
