+++ assets/js/AddCreditPage.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import axios from 'axios';
+import DollarInput from './DollarInput';
 
 
 axios.defaults.xsrfHeaderName = "X-CSRFTOKEN";
@@ -32,26 +33,8 @@ class AddCreditPage extends Component {
          });
     }
 
-    onAmountChange (amount) {
-        if (amount === undefined) {
-            return;
-        }
-
-        let amountStr = String(amount).replace('.', ''),
-            len = amountStr.length;
-        if (len === 1) {
-            amountStr = amountStr + '.00';
-        } else if (len === 2) {
-            amountStr = amountStr.slice(0, 1) + '.' + amountStr.slice(1, 2) + '0';
-        } else {
-            amountStr = amountStr.slice(0, len - 2) + '.' + amountStr.slice(len - 2, len);
-        }
-
-        if (isNaN(Number(amountStr))) {
-            this.setState({amount: ''});
-        } else {
-            this.setState({amount: Number(amountStr)});
-        }
+    updateAmount (amount) {
+        this.setState({amount: amount});
     }
 
     render () {
@@ -69,14 +52,7 @@ class AddCreditPage extends Component {
                 <div id="calculate" className="jumbotron row center-block">
                     <div className="total">
                         <h3 className="text-center">Amount to add:</h3>
-                        <input id="amount"
-                            className="numbers col-sm-offset-3 input-lg center-block text-center"
-                            type="number"
-                            min="0"
-                            step="0.25"
-                            value={this.state.amount}
-                            placeholder={0}
-                            onChange={(event) => this.onAmountChange(event.target.value)} /> 
+                        <DollarInput updateAmount={(amount) => this.updateAmount(amount)} /> 
                     </div>
                 </div>
                 <div>
